stl-telegram-bot/requirements.txt
python-telegram-bot==20.7
python-dotenv==1.0.0
fdb==2.0.2
httpx==0.25.2
pydantic==2.5.0
schedule==1.2.0